requirements.txt
Django>=2.2.4
pytesseract==0.2.7
celery==4.4.0
redis==3.3.11
Markdown==3.1.1
django-filter==2.2.0
djangorestframework==3.11.0
